+++ src/processors/normalize.ts
@@ -1,20 +1,22 @@
 import type { PipelineComponent } from "../core/types";
-import { IntentResult } from "../types";
+import { QirrelContext } from "../types";
 
 export const normalize: PipelineComponent = (
-  input: IntentResult,
-): IntentResult => {
-  input.tokens.forEach((token) => {
-    if (token.type === "word") {
-      token.value = token.value.toLowerCase();
-    } else if (token.type === "number") {
-      token.value = parseFloat(token.value).toString();
-    }
-  });
+  input: QirrelContext,
+): QirrelContext => {
+  if (input.data) {
+    input.data.tokens.forEach((token) => {
+      if (token.type === "word") {
+        token.value = token.value.toLowerCase();
+      } else if (token.type === "number") {
+        token.value = parseFloat(token.value).toString();
+      }
+    });
 
-  input.entities.forEach((entity) => {
-    entity.value = entity.value.toLowerCase();
-  });
+    input.data.entities.forEach((entity) => {
+      entity.value = entity.value.toLowerCase();
+    });
+  }
 
   return input;
 };